GC: Move plugins handling code to GamecubePlugins.c
Having the plugin handling code in GamecubeMain.cpp and gc_plugin.c is
problematic, because they both includes <libpcsxcore/plugin.h>, which
may define some symbols differently.
