+++ CMakeLists.txt
@@ -315,6 +315,7 @@ add_executable(wiisx
 	Gamecube/sysconf.c
 	Gamecube/xxhash.c
 	Gamecube/GamecubeMain.cpp
+	Gamecube/GamecubePlugins.c
 	Gamecube/vm/vm.c
 	Gamecube/vm/dsihandler.s
 	Gamecube/gc_input/controller-Classic.c
