+++ Gamecube/GamecubeMain.cpp
@@ -34,6 +34,7 @@
 # include <debug.h>
 #endif
 #include <libpcsxcore/psxcommon.h>
+#include <libpcsxcore/r3000a.h>
 #include "wiiSXconfig.h"
 #include "menu/MenuContext.h"
 extern "C" {
@@ -50,6 +51,8 @@ extern "C" {
 #endif
 #include "vm/vm.h"
 }
+
+#include <psemu_plugin_defs.h>
 
 #ifdef WII
 #include "MEM2.h"
@@ -336,18 +339,6 @@ void video_mode_init(GXRModeObj *videomode, u32 *fb1, u32 *fb2)
 	xfb[1] = fb2;
 }
 
-// Plugin structure
-extern "C" {
-#include "GamecubePlugins.h"
-PluginTable plugins[] =
-	{ PLUGIN_SLOT_0,
-	  PLUGIN_SLOT_1,
-	  PLUGIN_SLOT_2,
-	  PLUGIN_SLOT_3,
-	  PLUGIN_SLOT_4,
-	  PLUGIN_SLOT_5 };
-}
-
 int main(int argc, char *argv[]) 
 {
 	/* INITIALIZE */
@@ -618,6 +609,9 @@ void SysStartCPU() {
 	psxCpu->Execute();
 }
 
+extern "C" void ClosePlugins();
+extern "C" void ReleasePlugins();
+
 void SysClose() 
 {
 	psxShutdown();
@@ -660,27 +654,6 @@ void SysPrintf(const char *fmt, ...)
 #endif
 }
 
-void *SysLoadLibrary(const char *lib) 
-{
-	int i;
-	for(i=0; i<NUM_PLUGINS; i++)
-		if((plugins[i].lib != NULL) && (!strcmp(lib, plugins[i].lib)))
-			return (void*)i;
-	SysPrintf("SysLoadLibrary(%s) couldn't be found!\r\n", lib);
-	return NULL;
-}
-
-void *SysLoadSym(void *lib, const char *sym) 
-{
-	PluginTable* plugin = plugins + (int)lib;
-	int i;
-	for(i=0; i<plugin->numSyms; i++)
-		if(plugin->syms[i].sym && !strcmp(sym, plugin->syms[i].sym))
-			return plugin->syms[i].pntr;
-	SysPrintf("SysLoadSym(%s, %s) couldn't be found!\r\n", lib, sym);
-	return NULL;
-}
-
 int framesdone = 0;
 void SysUpdate() 
 {
@@ -689,8 +662,6 @@ void SysUpdate()
 
 void SysRunGui() {}
 void SysMessage(const char *fmt, ...) {}
-void SysCloseLibrary(void *lib) {}
-const char *SysLibError() {	return NULL; }
 
 void pl_frame_limit(void)
 {
